GH-1: Rename the poller package to services.
