+++ app/src/main/java/com/syntactec/subreddify/services/RedditPostsDeserializer.java
@@ -1,4 +1,4 @@
-package com.syntactec.subreddify.poller;
+package com.syntactec.subreddify.services;
 
 import com.google.gson.*;
 
@@ -6,7 +6,7 @@ import java.lang.reflect.Type;
 import java.util.ArrayList;
 import java.util.List;
 
-class RedditPostsDeserializer implements JsonDeserializer<List<RedditPost>> {
+public class RedditPostsDeserializer implements JsonDeserializer<List<RedditPost>> {
     @Override
     public List<RedditPost> deserialize(JsonElement jsonElement, Type type, JsonDeserializationContext context)
             throws JsonParseException {
