+++ app/src/main/java/com/syntactec/subreddify/services/RedditService.java
@@ -1,4 +1,4 @@
-package com.syntactec.subreddify.poller;
+package com.syntactec.subreddify.services;
 
 import retrofit2.Call;
 import retrofit2.http.GET;
